Change voting model to use an enum for values.
Remove the constant VALID_VALUES from the vote model and remove the
custom validation as well since they are no longer needed. Tests
were changed to reflect the the fact an ArgumentError is raised when an
invalid parameter is passed to create or update. The test for the quote
object's score function was also modified to use the enum. I18n was
utilized to implement the enum.