Change voting model to use an enum for values.
Remove the constant VALID_VALUES fromt eh vote model and remove the
custom validation as well since they are no longer needed.  Modify the
application controller helper to reflect the change to the enum.  Tests
were changed to reflect the the fact an ArgumentError is raised when an
invalid parameter is passed to create or update.  The test for the quote
object's score function was also modified to use the enum.