Remove extraneous helper method from quotes C.
A helper method had been used in the views before to be iterated through
and fill out the links for voting with the proper messages and values.
Since changing to an enum a hash stored in I18n is utilized instead so
this helper method is deprecated and no longer used.